@@ -36,12 +36,13 @@
     "lodash": "^4.17.21",
     "query-string": "^6.14.0",
     "rc-slider": "^8.6.13",
-    "react": "16.14.0",
+    "react": "18.0.0",
     "react-barcode-reader": "^0.0.2",
     "react-cropper": "^2.1.8",
     "react-datepicker": "^2.9.3",
-    "react-dom": "^16.8.2",
+    "react-dom": "18.0.0",
     "react-google-autocomplete": "^1.1.2",
+    "react-is": "18.0.0",
     "react-qr-reader": "^2.2.1",
     "react-signature-canvas": "^1.0.1",
     "reflect-metadata": "^0.1.13",
@@ -55,49 +56,52 @@
     "@babel/preset-env": "7.12.13",
     "@babel/preset-react": "7.12.13",
     "@babel/preset-typescript": "7.12.13",
-    "@nrwl/cli": "12.6.3",
-    "@nrwl/cypress": "12.6.3",
-    "@nrwl/eslint-plugin-nx": "12.6.3",
-    "@nrwl/jest": "12.6.3",
-    "@nrwl/linter": "12.6.3",
-    "@nrwl/react": "^12.6.3",
-    "@nrwl/tao": "12.6.3",
-    "@nrwl/web": "12.6.3",
-    "@nrwl/workspace": "12.6.3",
-    "@testing-library/react": "11.2.6",
+    "@nrwl/cli": "15.2.1",
+    "@nrwl/cypress": "15.2.1",
+    "@nrwl/eslint-plugin-nx": "15.2.1",
+    "@nrwl/jest": "15.2.1",
+    "@nrwl/linter": "15.2.1",
+    "@nrwl/nx-cloud": "15.0.2",
+    "@nrwl/react": "15.2.1",
+    "@nrwl/web": "15.2.1",
+    "@nrwl/workspace": "15.2.1",
+    "@testing-library/react": "13.4.0",
     "@types/bootstrap": "^4.3.1",
     "@types/googlemaps": "^3.43.3",
-    "@types/jest": "26.0.24",
+    "@types/jest": "28.1.8",
     "@types/jquery": "^3.2.1",
     "@types/lodash": "^4.14.177",
-    "@types/node": "14.14.33",
+    "@types/node": "18.7.18",
     "@types/rc-slider": "^8.6.4",
-    "@types/react": "16.8.3",
+    "@types/react": "18.0.20",
     "@types/react-datepicker": "^2.3.0",
-    "@types/react-dom": "17.0.3",
+    "@types/react-dom": "18.0.6",
     "@types/react-qr-reader": "^2.1.4",
     "@types/react-signature-canvas": "^1.0.2",
     "@types/ts-nameof": "^4.2.1",
-    "@typescript-eslint/eslint-plugin": "~4.28.3",
-    "@typescript-eslint/parser": "~4.28.3",
-    "babel-jest": "27.0.6",
+    "@typescript-eslint/eslint-plugin": "5.44.0",
+    "@typescript-eslint/parser": "5.44.0",
+    "babel-jest": "28.1.3",
     "babel-plugin-ts-nameof": "^4.2.1",
     "cypress": "^7.3.0",
-    "dotenv": "~10.0.0",
-    "eslint": "7.22.0",
+    "dotenv": "10.0.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.23.1",
-    "eslint-plugin-react-hooks": "4.2.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-react": "7.31.8",
+    "eslint-plugin-react-hooks": "4.6.0",
     "file-loader": "^6.2.0",
-    "jest": "27.0.3",
-    "prettier": "^2.3.1",
-    "ts-jest": "27.0.3",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "^29.3.1",
+    "nx": "15.2.1",
+    "prettier": "2.7.1",
+    "react-test-renderer": "18.0.0",
+    "ts-jest": "28.0.8",
     "ts-nameof": "^5.0.0",
-    "ts-node": "~9.1.1",
-    "typescript": "3.8.2"
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4"
   },
   "peerDependencies": {
     "jquery": "^3.2.1"
