@@ -1,8 +1,9 @@
-module.exports = {
+/* eslint-disable */
+export default {
   displayName: 'WeAre.ReApptor.React.Components',
   preset: '../jest.preset.js',
   transform: {
-    '^.+\\.[tj]sx?$': 'babel-jest',
+    '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nrwl/react/babel'] }],
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
   coverageDirectory: '../coverage/libs/WeAre.ReApptor.React.Components'
