@@ -1,18 +1,17 @@
 import React from "react";
-import {BaseComponent} from "@weare/reapptor-react-common";
+import {BaseComponent, IBaseContainerComponentProps} from "@weare/reapptor-react-common";
 
 import styles from "./ToolbarRow.module.scss";
 import Inline, {JustifyContent} from "../../Layout.Inline/Inline";
 
-interface IToolbarRowProps {
+interface IToolbarRowProps extends IBaseContainerComponentProps {
     justify?: JustifyContent;
-    className?: string;
 }
 
 export default class ToolbarRow extends BaseComponent<IToolbarRowProps> {
     public render() {
         return (
-            <Inline className={this.css(styles.toolbarRow, this.props.className)} justify={this.props.justify}>
+            <Inline id={this.id} className={this.css(styles.toolbarRow, this.props.className)} justify={this.props.justify}>
                 {this.props.children}
             </Inline>
         )
